#4920 refactor code and use the newest price lsit version when import initial IFA files
#4920 Improve creating product prices when importing IFA products
